sched/cass: Achieve effects similar to prefer_high_cap
Cass will select the CPU with lower relative utilization so that the
relative utilization of all CPUs is similar. Because the utilization of
some CPUs has been scaled, CPUs with higher capacity are not selected
for some important tasks, resulting in insufficient smoothness on
some SOCs. Let some tasks select CPUs based on utilization and
capacity to optimize the experience.